Refactor goal channel projection read model
